Correctly check error returned by dwarf_get_scopevar.
frysk-core/frysk/debuginfo/ChangeLog
2008-06-15  Andrew Cagney  <cagney@redhat.com>

	* gen-type-expect-tests.py (open_file): Use assert, not if/println
	to detect and report bad values.

frysk-sys/lib/dwfl/ChangeLog
2008-06-15  Andrew Cagney  <cagney@redhat.com>

	* DwarfDie.java (get_scopevar): Return an int.
	* jni/DwarfDie.cxx (DwarfDie::get_scopevar): Only examine the die
	when its valid.
	* cni/DwarfDie.cxx (DwarfDie::get_scopevar): Ditto.

+++ frysk-sys/lib/dwfl/DwarfDie.java
@@ -476,7 +476,8 @@ public class DwarfDie {
 
     private native long[] get_scopes_die ();
 
-    private native long get_scopevar (long[] die_scope, long[] scopes, String variable);
+    private native int get_scopevar(long[] die_scope, long[] scopes,
+				    String variable);
 
     private native long get_scopevar_names (long[] scopes, String variable);
